Run specs for the installed gem
- extract shared specs for library and installed gem
- gem requires html validator classes
- rubocop ignores shared_specs
