+++ lib/bashtetikn.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 require_relative "bashtetikn/version"
+require_relative "bashtetikn/html_validator_from_w3c"
 
 module Bashtetikn
   class Error < StandardError; end
