--- spec/cassettes/Bashtetikn_HtmlValidatorFromW3C/_validate_text/for_empty_text/returns_2_errors.yml
@@ -1,57 +1,0 @@
----
-http_interactions:
-- request:
-    method: post
-    uri: https://validator.w3.org/nu/?out=json&parser=html&showsource=yes
-    body:
-      encoding: UTF-8
-      string: ''
-    headers:
-      Content-Type:
-      - text/html; charset=utf-8
-      Accept-Encoding:
-      - gzip;q=1.0,deflate;q=0.6,identity;q=0.3
-      Accept:
-      - "*/*"
-      User-Agent:
-      - Ruby
-  response:
-    status:
-      code: 200
-      message: ''
-    headers:
-      Date:
-      - Thu, 14 Oct 2021 03:00:00 GMT
-      Accept-Encoding:
-      - gzip
-      Access-Control-Allow-Origin:
-      - "*"
-      Access-Control-Allow-Headers:
-      - content-type
-      Expires:
-      - Thu, 01 Jan 1970 00:00:00 GMT
-      Cache-Control:
-      - no-cache
-      Content-Type:
-      - application/json;charset=utf-8
-      Vary:
-      - Accept-Encoding, User-Agent
-      Server:
-      - Jetty(9.4.43.v20210629)
-      Strict-Transport-Security:
-      - max-age=15552015; preload
-      Public-Key-Pins:
-      - pin-sha256="cN0QSpPIkuwpT6iP2YjEo1bEwGpH/yiUn6yhdy+HNto="; pin-sha256="WGJkyYjx1QMdMe0UqlyOKXtydPDVrk7sl2fV+nNm1r4=";
-        pin-sha256="LrKdTxZLRTvyHM4/atX2nquX9BeHRZMCxg3cf4rhc2I="; max-age=864000
-      X-Frame-Options:
-      - deny
-      X-Xss-Protection:
-      - 1; mode=block
-      Transfer-Encoding:
-      - chunked
-    body:
-      encoding: ASCII-8BIT
-      string: !binary |-
-        eyJtZXNzYWdlcyI6W3sidHlwZSI6ImVycm9yIiwibWVzc2FnZSI6IkVuZCBvZiBmaWxlIHNlZW4gd2l0aG91dCBzZWVpbmcgYSBkb2N0eXBlIGZpcnN0LiBFeHBlY3RlZCDigJw8IURPQ1RZUEUgaHRtbD7igJ0uIn0seyJ0eXBlIjoiZXJyb3IiLCJtZXNzYWdlIjoiRWxlbWVudCDigJxoZWFk4oCdIGlzIG1pc3NpbmcgYSByZXF1aXJlZCBpbnN0YW5jZSBvZiBjaGlsZCBlbGVtZW50IOKAnHRpdGxl4oCdLiJ9LHsidHlwZSI6ImluZm8iLCJzdWJUeXBlIjoid2FybmluZyIsIm1lc3NhZ2UiOiJDb25zaWRlciBhZGRpbmcgYSDigJxsYW5n4oCdIGF0dHJpYnV0ZSB0byB0aGUg4oCcaHRtbOKAnSBzdGFydCB0YWcgdG8gZGVjbGFyZSB0aGUgbGFuZ3VhZ2Ugb2YgdGhpcyBkb2N1bWVudC4ifV0sInNvdXJjZSI6eyJ0eXBlIjoidGV4dC9odG1sIiwiZW5jb2RpbmciOiJVVEYtOCIsImNvZGUiOiIifX0K
-  recorded_at: Thu, 14 Oct 2021 03:00:00 GMT
-recorded_with: VCR 6.0.0
