--- spec/cassettes/Bashtetikn_HtmlValidatorFromW3C/_validate_text/for_invalid_html/returns_2_errors.yml
@@ -1,67 +1,0 @@
----
-http_interactions:
-- request:
-    method: post
-    uri: https://validator.w3.org/nu/?out=json&parser=html&showsource=yes
-    body:
-      encoding: UTF-8
-      string: |2
-                <!DOCTYPE html>
-                <html lang="en-ca">
-                  <head>
-                    <title>HTML 5 Example</title>
-                  </head>
-                  <body>
-                    <section>Unclosed element
-                    <p>This is a sample HTML 5 document with an unclosed element.</p>
-                  </body>
-                </html>
-    headers:
-      Content-Type:
-      - text/html; charset=utf-8
-      Accept-Encoding:
-      - gzip;q=1.0,deflate;q=0.6,identity;q=0.3
-      Accept:
-      - "*/*"
-      User-Agent:
-      - Ruby
-  response:
-    status:
-      code: 200
-      message: ''
-    headers:
-      Date:
-      - Thu, 14 Oct 2021 03:00:00 GMT
-      Accept-Encoding:
-      - gzip
-      Access-Control-Allow-Origin:
-      - "*"
-      Access-Control-Allow-Headers:
-      - content-type
-      Expires:
-      - Thu, 01 Jan 1970 00:00:00 GMT
-      Cache-Control:
-      - no-cache
-      Content-Type:
-      - application/json;charset=utf-8
-      Vary:
-      - Accept-Encoding, User-Agent
-      Server:
-      - Jetty(9.4.43.v20210629)
-      Strict-Transport-Security:
-      - max-age=15552015; preload
-      Public-Key-Pins:
-      - pin-sha256="cN0QSpPIkuwpT6iP2YjEo1bEwGpH/yiUn6yhdy+HNto="; pin-sha256="WGJkyYjx1QMdMe0UqlyOKXtydPDVrk7sl2fV+nNm1r4=";
-        pin-sha256="LrKdTxZLRTvyHM4/atX2nquX9BeHRZMCxg3cf4rhc2I="; max-age=864000
-      X-Frame-Options:
-      - deny
-      X-Xss-Protection:
-      - 1; mode=block
-      Transfer-Encoding:
-      - chunked
-    body:
-      encoding: ASCII-8BIT
-      string: !binary |-
-        eyJtZXNzYWdlcyI6W3sidHlwZSI6ImVycm9yIiwibGFzdExpbmUiOjksImxhc3RDb2x1bW4iOjE3LCJmaXJzdENvbHVtbiI6MTEsIm1lc3NhZ2UiOiJFbmQgdGFnIGZvciAg4oCcYm9keeKAnSBzZWVuLCBidXQgdGhlcmUgd2VyZSB1bmNsb3NlZCBlbGVtZW50cy4iLCJleHRyYWN0IjoiICAgICAgICAgIDwvYm9keT5cbiAgICAgIiwiaGlsaXRlU3RhcnQiOjEwLCJoaWxpdGVMZW5ndGgiOjd9LHsidHlwZSI6ImVycm9yIiwibGFzdExpbmUiOjcsImxhc3RDb2x1bW4iOjIxLCJmaXJzdENvbHVtbiI6MTMsIm1lc3NhZ2UiOiJVbmNsb3NlZCBlbGVtZW50IOKAnHNlY3Rpb27igJ0uIiwiZXh0cmFjdCI6IiAgICAgICAgICA8c2VjdGlvbj5VbmNsb3MiLCJoaWxpdGVTdGFydCI6MTAsImhpbGl0ZUxlbmd0aCI6OX0seyJ0eXBlIjoiaW5mbyIsImxhc3RMaW5lIjo3LCJsYXN0Q29sdW1uIjoyMSwiZmlyc3RDb2x1bW4iOjEzLCJzdWJUeXBlIjoid2FybmluZyIsIm1lc3NhZ2UiOiJTZWN0aW9uIGxhY2tzIGhlYWRpbmcuIENvbnNpZGVyIHVzaW5nIOKAnGgy4oCdLeKAnGg24oCdIGVsZW1lbnRzIHRvIGFkZCBpZGVudGlmeWluZyBoZWFkaW5ncyB0byBhbGwgc2VjdGlvbnMuIiwiZXh0cmFjdCI6IiAgICAgICAgICA8c2VjdGlvbj5VbmNsb3MiLCJoaWxpdGVTdGFydCI6MTAsImhpbGl0ZUxlbmd0aCI6OX1dLCJzb3VyY2UiOnsidHlwZSI6InRleHQvaHRtbCIsImVuY29kaW5nIjoiVVRGLTgiLCJjb2RlIjoiICAgICAgICA8IURPQ1RZUEUgaHRtbD5cbiAgICAgICAgPGh0bWwgbGFuZz1cImVuLWNhXCI+XG4gICAgICAgICAgPGhlYWQ+XG4gICAgICAgICAgICA8dGl0bGU+SFRNTCA1IEV4YW1wbGU8L3RpdGxlPlxuICAgICAgICAgIDwvaGVhZD5cbiAgICAgICAgICA8Ym9keT5cbiAgICAgICAgICAgIDxzZWN0aW9uPlVuY2xvc2VkIGVsZW1lbnRcbiAgICAgICAgICAgIDxwPlRoaXMgaXMgYSBzYW1wbGUgSFRNTCA1IGRvY3VtZW50IHdpdGggYW4gdW5jbG9zZWQgZWxlbWVudC48L3A+XG4gICAgICAgICAgPC9ib2R5PlxuICAgICAgICA8L2h0bWw+In19Cg==
-  recorded_at: Thu, 14 Oct 2021 03:00:00 GMT
-recorded_with: VCR 6.0.0
