+++ spec/cassettes/Bashtetikn_HtmlValidatorFromW3C/it_should_behave_like_an_html_validator/_validate_text/for_valid_html/returns_0_errors.yml
@@ -1,0 +1,76 @@
+---
+http_interactions:
+- request:
+    method: post
+    uri: https://validator.w3.org/nu/?out=json&parser=html&showsource=yes
+    body:
+      encoding: UTF-8
+      string: |2
+                <!DOCTYPE html>
+                <html lang="en-ca">
+                  <head>
+                    <title>HTML 5 Example</title>
+                  </head>
+                  <body>
+                    <p>This is a sample HTML 5 document.</p>
+                    <h1>Example of paragraphs</h1>
+                    This is the <em>first</em> paragraph in this example.
+                    <p>This is the second.</p>
+                    <p>Test<br>test</p>
+                  </body>
+                </html>
+    headers:
+      Content-Type:
+      - text/html; charset=utf-8
+      Accept-Encoding:
+      - gzip;q=1.0,deflate;q=0.6,identity;q=0.3
+      Accept:
+      - "*/*"
+      User-Agent:
+      - Ruby
+  response:
+    status:
+      code: 200
+      message: ''
+    headers:
+      Date:
+      - Thu, 21 Oct 2021 01:26:37 GMT
+      Accept-Encoding:
+      - gzip
+      Access-Control-Allow-Origin:
+      - "*"
+      Access-Control-Allow-Headers:
+      - content-type
+      Expires:
+      - Thu, 01 Jan 1970 00:00:00 GMT
+      Cache-Control:
+      - no-cache
+      Content-Type:
+      - application/json;charset=utf-8
+      Vary:
+      - Accept-Encoding, User-Agent
+      Server:
+      - Jetty(9.4.43.v20210629)
+      Strict-Transport-Security:
+      - max-age=15552015; preload
+      Public-Key-Pins:
+      - pin-sha256="cN0QSpPIkuwpT6iP2YjEo1bEwGpH/yiUn6yhdy+HNto="; pin-sha256="WGJkyYjx1QMdMe0UqlyOKXtydPDVrk7sl2fV+nNm1r4=";
+        pin-sha256="LrKdTxZLRTvyHM4/atX2nquX9BeHRZMCxg3cf4rhc2I="; max-age=864000
+      X-Frame-Options:
+      - deny
+      X-Xss-Protection:
+      - 1; mode=block
+      Transfer-Encoding:
+      - chunked
+    body:
+      encoding: ASCII-8BIT
+      string: '{"messages":[],"source":{"type":"text/html","encoding":"UTF-8","code":"        <!DOCTYPE
+        html>\n        <html lang=\"en-ca\">\n          <head>\n            <title>HTML
+        5 Example</title>\n          </head>\n          <body>\n            <p>This
+        is a sample HTML 5 document.</p>\n            <h1>Example of paragraphs</h1>\n            This
+        is the <em>first</em> paragraph in this example.\n            <p>This is the
+        second.</p>\n            <p>Test<br>test</p>\n          </body>\n        </html>"}}
+
+        '
+  recorded_at: Thu, 21 Oct 2021 01:26:37 GMT
+recorded_with: VCR 6.0.0
