+++ spec/cassettes/The_installed_gem_0_1_0_/Bashtetikn_HtmlValidatorFromW3C/it_should_behave_like_an_html_validator/_validate_uri/_a_valid_page_/should_be_valid.yml
@@ -1,0 +1,61 @@
+---
+http_interactions:
+- request:
+    method: get
+    uri: https://validator.w3.org/nu/?doc=https://w3c-validators.github.io/w3c_validators/valid_html5.html&out=json&parser=html&showsource=yes
+    body:
+      encoding: US-ASCII
+      string: ''
+    headers:
+      Accept-Encoding:
+      - gzip;q=1.0,deflate;q=0.6,identity;q=0.3
+      Accept:
+      - "*/*"
+      User-Agent:
+      - Ruby
+  response:
+    status:
+      code: 200
+      message: ''
+    headers:
+      Date:
+      - Thu, 21 Oct 2021 01:38:37 GMT
+      Accept-Encoding:
+      - gzip
+      Access-Control-Allow-Origin:
+      - "*"
+      Access-Control-Allow-Headers:
+      - content-type
+      Expires:
+      - Thu, 01 Jan 1970 00:00:00 GMT
+      Cache-Control:
+      - no-cache
+      Content-Type:
+      - application/json;charset=utf-8
+      Vary:
+      - Accept-Encoding, User-Agent
+      Server:
+      - Jetty(9.4.43.v20210629)
+      Strict-Transport-Security:
+      - max-age=15552015; preload
+      Public-Key-Pins:
+      - pin-sha256="cN0QSpPIkuwpT6iP2YjEo1bEwGpH/yiUn6yhdy+HNto="; pin-sha256="WGJkyYjx1QMdMe0UqlyOKXtydPDVrk7sl2fV+nNm1r4=";
+        pin-sha256="LrKdTxZLRTvyHM4/atX2nquX9BeHRZMCxg3cf4rhc2I="; max-age=864000
+      X-Frame-Options:
+      - deny
+      X-Xss-Protection:
+      - 1; mode=block
+      Transfer-Encoding:
+      - chunked
+    body:
+      encoding: ASCII-8BIT
+      string: '{"url":"https://w3c-validators.github.io/w3c_validators/valid_html5.html","messages":[],"source":{"type":"text/html","encoding":"utf-8","code":"<!DOCTYPE
+        html>\n<html lang=\"en-ca\">\n  <head>\n\t<meta http-equiv=\"Content-type\"
+        content=\"text/html; charset=utf-8\">\n    <title>HTML 5 Example</title>\n  </head>\n  <body>\n    <p>This
+        is a sample HTML 5 document.</p>\n    <section>\n\t  <h1>Example of paragraphs</h1>\n\t  This
+        is the <em>first</em> paragraph in this example.\n\t  <p>This is the second.</p>\n\t  <!--
+        This is not a paragraph. -->\n\t</section>\n  </body>\n</html>"}}
+
+        '
+  recorded_at: Thu, 21 Oct 2021 01:38:37 GMT
+recorded_with: VCR 6.0.0
